+++ src/util/mod.rs
@@ -2,3 +2,4 @@
 // SPDX-License-Identifier: GPL-2.0-only
 
 pub mod bin_version;
+pub mod cpu_arch;